+++ src/fairseq2/composition/lib.py
@@ -50,6 +50,7 @@ from fairseq2.model_checkpoint import (
 )
 from fairseq2.models.hub import GlobalModelLoader
 from fairseq2.models.llama import LLaMACheckpointLoader
+from fairseq2.models.llama4.sharder import MoESharder
 from fairseq2.runtime.dependency import DependencyContainer, DependencyResolver
 from fairseq2.sharder import (
     EmbeddingSharder,
@@ -148,6 +149,7 @@ def _register_library(container: DependencyContainer) -> None:
 
     container.collection.register_type(ModuleSharder, EmbeddingSharder)
     container.collection.register_type(ModuleSharder, LinearSharder)
+    container.collection.register_type(ModuleSharder, MoESharder)
 
     container.collection.register_type(ModelCheckpointLoader, BasicModelCheckpointLoader)
     container.collection.register_type(ModelCheckpointLoader, NativeModelCheckpointLoader)